+++ src/Filament/Resources/VacancyResource/Pages/CreateVacancy.php
@@ -3,7 +3,6 @@
 namespace Dashed\DashedVacancies\Filament\Resources\VacancyResource\Pages;
 
 use Filament\Resources\Pages\CreateRecord;
-use Dashed\DashedVacancies\Classes\VacancyForms;
 use Dashed\DashedVacancies\Filament\Resources\VacancyResource;
 use Dashed\DashedCore\Filament\Concerns\HasCreatableCMSActions;
 
@@ -16,20 +15,5 @@ class CreateVacancy extends CreateRecord
     protected function getActions(): array
     {
         return self::CMSActions();
-    }
-
-    protected function mutateFormDataBeforeCreate(array $data): array
-    {
-        $autoCreate = (bool) ($data['auto_create_form'] ?? false);
-        unset($data['auto_create_form']);
-
-        if ($autoCreate && empty($data['form_id']) && class_exists(\Dashed\DashedForms\Models\Form::class)) {
-            $form = VacancyForms::createApplicationForm($data['name'] ?? null);
-            if ($form) {
-                $data['form_id'] = $form->id;
-            }
-        }
-
-        return $data;
     }
 }